pyproject.toml
[project]
name = "helix-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cerebras-cloud-sdk>=1.50.1",
    "fastapi>=0.118.0",
    "google-cloud-firestore>=2.21.0",
    "markitdown[all]>=0.1.3",
    "openai>=2.0.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
